A-More-OOP: Adding more abstraction through types, storage, ui and utils class
