+++ src/main/java/Barbie.java
@@ -1,15 +1,18 @@
+import types.Deadlines;
+import types.Party;
+import types.Task;
+import types.Todo;
+import exceptions.*;
+
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.*;
-import java.io.IOException;
 
 
 public class Barbie {
-
 
     enum Command {
         MARK,
@@ -25,29 +28,14 @@ public class Barbie {
 
     public static void main(String[] args) {
         // CONSTANTS
-        String line = "______________________________\n";
         Scanner scanner = new Scanner(System.in);
-        String logo = " ____             _\n"
-                + "|  _ \\           | |\n"
-                + "| |_| |_____,_ ,_| |,___  _  ___\n"
-                + "|  _ /|  _  | ` _|  __\\ \\ |/ _  \\\n"
-                + "| |_| | |_| | |  | |__/ /| |  ___/\n"
-                + "|____/ \\__,_|_|  |_|\\__/ |_|\\___/\n";
-        ArrayList<Task> list = getLastList();
+        ArrayList<Task> list = Storage.getLastList();
         Path path = Paths.get("barbie.txt");
         int indexNumber = list.size(); // Starting from 1 reduces the need to subtract and add 1 for usability.
 
 
         // Intro
-            System.out.println(line
-                    + "Hi Barbie! Hi Ken!\n"
-                    + "\nI'm\n"
-                    + logo
-                    + "\n\nThis is the list of things you have today!");
-            getDateList(LocalDate.now(), list).forEach(System.out::println);
-            System.out.println("\nWhat can I do for you?\n" + line);
-
-        System.out.println("[you]:");
+        Ui.intro(Utils.getDateList(LocalDate.now(), list));
 
             loop:
             while (true) {
@@ -57,8 +45,7 @@ public class Barbie {
                     String[] parts = input.split(" ", 2);
                     Command command = Command.valueOf(parts[0].toUpperCase());
 
-                    System.out.println("\t" + line
-                            + "\t [barbie]:\n");
+                    Ui.barbie();
 
                     switch (command) {
                         case MARK:
@@ -77,36 +64,32 @@ public class Barbie {
                                 case MARK:
                                     // Editing variables
                                     list.get(taskNumber).mark();
-                                    changeLineStatus(path, "1", taskNumber);
+                                    Storage.changeLineStatus(path, "1", taskNumber);
 
                                     // Output
-                                    System.out.println("\t Nice! I've marked this task as done:\n"
-                                            + "\t " + list.get(taskNumber) + "\n"
-                                            + "\t" + line);
+                                    Ui.mark(list.get(taskNumber));
                                     break;
 
                                 case UNMARK:
                                     // Editing variables
                                     taskNumber = Integer.parseInt(desc);
                                     list.get(taskNumber).unmark();
-                                    changeLineStatus(path, "2", taskNumber);
+                                    Storage.changeLineStatus(path, "2", taskNumber);
 
                                     // Output
-                                    System.out.println("\t Alright! I've marked this task as not done yet:\n"
-                                            + "\t " + list.get(taskNumber) + "\n"
-                                            + "\t" + line);
+                                    Ui.unmark(list.get(taskNumber));
+
                                     break;
 
                                 case DEL:
                                     // Editing variables
                                     list.remove(taskNumber);
                                     indexNumber -= 1;
-                                    deleteLine(path, taskNumber);
+                                    Storage.deleteLine(path, taskNumber);
 
                                     // Output
-                                    System.out.println("\t Deletion success! I've deleted this task off your list.");
+                                    Ui.del();
                                     break;
-
                             }
                             break;
 
@@ -118,15 +101,17 @@ public class Barbie {
                             }
                             desc = parts[1];
                             String[] parts2 = parts[1].split("/");
+
                             switch (command) {
                                 case DEADLINE:
                                     if (parts2.length < 2) {
                                         throw new BarbieNoDeadlineException();
                                     }
+
                                     desc = parts2[0];
                                     LocalDate by = LocalDate.parse(parts2[1]);
                                     list.add(indexNumber, new Deadlines(desc, by));
-                                    addToList(path, "D", desc, by);
+                                    Storage.addToList(path, "D", desc, by);
 
                                     break;
 
@@ -134,28 +119,29 @@ public class Barbie {
                                     if (parts2.length < 3) {
                                         throw new BarbieNoTimingException();
                                     }
+
                                     desc = parts2[0];
                                     LocalDate from = LocalDate.parse(parts2[1].strip());
                                     LocalDate to = LocalDate.parse(parts2[2].strip());
+
                                     list.add(indexNumber, new Party(desc, from, to));
-                                    addToList(path, "P", desc, from, to);
+                                    Storage.addToList(path, "P", desc, from, to);
                                     break;
 
                                 default:
                                     list.add(indexNumber, new Todo(desc));
-                                    addToList(path, "T", desc);
+                                    Storage.addToList(path, "T", desc);
                                     break;
 
                             }
 
-                            System.out.println("\tGot you barbie! I've added this task to your Barbie list:\n"
-                                    + "\t " + list.get(indexNumber));
+                            Ui.taskAdded(list.get(indexNumber));
                             indexNumber ++;
                             break;
 
                         case LIST:
                             // No variables to edit, only output (refer to listTasks func)
-                            listTasks(list, indexNumber);
+                            Ui.listTasks(list, indexNumber);
                             break;
 
                         case BYE:
@@ -184,147 +170,14 @@ public class Barbie {
 
                 }
 
-                System.out.println("\t" + line);
                 System.out.println("[you]:");
 
             }
 
             // Exit
-            System.out.println(line + "Bye Barbie! Bye Ken!\n" + line);
+            Ui.exit();
 
     }
 
-    protected static void listTasks(ArrayList<Task> list,int indexNumber) {
-        if (indexNumber == 0) {
-            System.out.println("Hmm.. your list looks empty. To add items, use the 'todo', 'deadline' or 'party' commands!");
-        }
-        // "list" command
-        for (int i = 0; i < indexNumber; i++) {
-            int itemNumber = i + 1;
-            System.out.println("\t" + itemNumber + ". " + list.get(i));
-        }
-    }
 
-    protected static ArrayList<Task> getLastList() {
-        ArrayList<Task> finalList = new ArrayList<>();
-
-        try {
-            Path path = Paths.get("barbie.txt");
-
-            if (!Files.exists(path)) {
-                Files.createFile(path);
-                System.out.println("[A new list created for current user]");
-            } else {
-                System.out.println("-------------------------------------------------");
-                System.out.println("[A current list is being used for current user]");
-                Files.readAllLines(path).forEach(x -> {
-
-                    String[] taskParts = x.split(",");
-                    String taskType = taskParts[0];
-                    String taskStatus = taskParts[1];
-                    String desc = taskParts[2];
-                    Task task;
-
-                    if (Objects.equals(taskType, "T")) {
-                        task = new Todo(desc);
-                    } else if (Objects.equals(taskType, "D")) {
-                        task = new Deadlines(desc, LocalDate.parse(taskParts[3]));
-                    } else if (Objects.equals(taskType, "P")) {
-                        task = new Party(desc, LocalDate.parse(taskParts[3]), LocalDate.parse(taskParts[4]));
-                    } else {
-                        task = new Task(desc);
-                    }
-                    if (Integer.parseInt( taskStatus) == 1) { task.mark(); }
-                    finalList.add(task);
-                    System.out.println(finalList);
-                });
-
-            }
-
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return finalList;
-
-    }
-
-    protected static void deleteLine(Path path, int lineToDelete) {
-        try {
-            List<String> lines = Files.readAllLines(path);
-
-            if (lineToDelete >= 0 && lineToDelete < lines.size()) {
-                lines.remove(lineToDelete);
-                Files.write(path, lines);
-            } else {
-                throw new IllegalArgumentException("Invalid line number to delete.");
-            }
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
-    protected static void addToList(Path path, String type, String desc) {
-        String line = type + "," + 0 + "," + desc + "\n";
-        try {
-            Files.write(path, line.getBytes(), StandardOpenOption.APPEND);
-        } catch (IOException e) {
-            System.out.println(e);
-        }
-    }
-
-    protected static void addToList(Path path, String type, String desc, LocalDate deadline) {
-        String line = type + "," + 0 + "," + desc + "," + deadline + "\n";
-        try {
-            Files.write(path, line.getBytes(), StandardOpenOption.APPEND);
-        } catch (IOException e) {
-            System.out.println(e);
-        }
-    }
-
-    protected static void addToList(Path path, String type, String desc, LocalDate from, LocalDate to) {
-        String line = type + "," + 0 + "," + desc + "," + from + "," + to + "\n";
-        try {
-            Files.write(path, line.getBytes(), StandardOpenOption.APPEND);
-        } catch (IOException e) {
-            System.out.println(e);
-        }
-    }
-
-    protected static void changeLineStatus(Path path, String status, int lineToChange) {
-        try {
-            List<String> lines = Files.readAllLines(path);
-
-            if (lineToChange >= 0 && lineToChange < lines.size()) {
-                String[] newContent = lines.get(lineToChange).split(",");
-                newContent[1] = status;
-                lines.set(lineToChange, Arrays.stream(newContent).reduce("", (x, acc) -> x + acc));
-                Files.write(path, lines);
-            } else {
-                throw new IllegalArgumentException("Invalid line number to change.");
-            }
-        } catch (IOException e) {
-            System.out.println(e);
-        }
-    }
-
-    protected static ArrayList<Task> getDateList(LocalDate date, ArrayList<Task> lastList) {
-        ArrayList<Task> thisDatesList = new ArrayList<>();
-
-        try {
-            lastList.forEach(x -> {
-                if (x instanceof Deadlines) {
-                    Deadlines y = (Deadlines) x;
-                    if (y.isToday(date)) { thisDatesList.add(y); }
-                } else if (x instanceof Party) {
-                    Party y = (Party) x;
-                    if (y.isToday(date)) { thisDatesList.add(y); }
-                }
-            });
-
-        } catch (Exception e) {
-            System.out.println(e);
-        }
-        return thisDatesList;
-    }
 }
