+++ src/main/java/exceptions/BarbieException.java
@@ -1,3 +1,5 @@
+package exceptions;
+
 public class BarbieException extends Exception{
     public BarbieException(String message) {
         super(message);
