+++ src/main/java/exceptions/BarbieTaskNumberException.java
@@ -1,4 +1,8 @@
-public class BarbieTaskNumberException extends BarbieException{
+package exceptions;
+
+import exceptions.BarbieException;
+
+public class BarbieTaskNumberException extends BarbieException {
     public BarbieTaskNumberException() {
         super("Task to mark or unmark or del is not provided as a digit!\n"
                 + "\n" + "use the list command to see the digit of your task and make sure to give \n"
