+++ src/main/java/types/Deadlines.java
@@ -1,3 +1,5 @@
+package types;
+
 import java.time.LocalDate;
 import java.util.Objects;
