+++ src/main/java/types/Party.java
@@ -1,3 +1,5 @@
+package types;
+
 import java.time.LocalDate;
 import java.util.Objects;
